Clean up cluster stats mechanism
