+++ src/toil/provisioners/clusterScaler.py
@@ -14,10 +14,13 @@
 
 from __future__ import absolute_import
 
+import json
 import logging
+import os
 from collections import deque
 from threading import Lock
 
+import time
 from bd2k.util.exceptions import require
 from bd2k.util.retry import retry
 from bd2k.util.threading import ExceptionalThread
@@ -314,9 +317,13 @@ class ScalerThread(ExceptionalThread):
         else:
             self.totalNodes = 0
         logger.info('Starting with %s %s(s) in the cluster.', self.totalNodes, self.nodeTypeString)
-        
+
+        self.stats = None
         if scaler.config.clusterStats:
-            self.scaler.provisioner.startStats(preemptable=preemptable)
+            self.stats = ClusterStats(self.scaler.leader.config.clusterStats,
+                                      self.scaler.leader.batchSystem,
+                                      self.scaler.provisioner.clusterName)
+            self.stats.startStats(preemptable=preemptable)
 
     def tryRun(self):
         global _preemptableNodeDeficit
@@ -424,7 +431,8 @@ class ScalerThread(ExceptionalThread):
                         else:
                             _preemptableNodeDeficit = 0
 
-                self.scaler.provisioner.checkStats()
+                if self.stats:
+                    self.stats.checkStats()
                     
         self.shutDown(preemptable=self.preemptable)
         logger.info('Scaler exited normally.')
@@ -611,8 +619,7 @@ class ScalerThread(ExceptionalThread):
         return nodeToInfo
 
     def shutDown(self, preemptable):
-        # only shutdown the stats threads once
-        self.scaler.provisioner._shutDownStats()
+        self.stats.shutDownStats()
         logger.debug('Forcing provisioner to reduce cluster size to zero.')
         totalNodes = self.setNodeCount(numNodes=0, preemptable=preemptable, force=True)
         if totalNodes > len(self.scaler.provisioner.getStaticNodes(preemptable)):  # ignore static nodes
@@ -621,3 +628,82 @@ class ScalerThread(ExceptionalThread):
                                )
         elif totalNodes < len(self.scaler.provisioner.getStaticNodes(preemptable)):  # ignore static nodes
             raise RuntimeError('Provisioner incorrectly terminated statically provisioned nodes.')
+
+
+class ClusterStats(object):
+
+    def __init__(self, path, batchSystem, clusterName):
+        self.stats = {}
+        self.statsThreads = []
+        self.statsPath = path
+        self.stop = False
+        self.clusterName = clusterName
+        self.batchSystem = batchSystem
+        self.scaleable = isinstance(self.batchSystem, AbstractScalableBatchSystem) if batchSystem else False
+
+    def shutDownStats(self):
+        if self.stop:
+            return
+        def getFileName():
+            extension = '.json'
+            file = '%s-stats' % self.clusterName
+            counter = 0
+            while True:
+                suffix = str(counter).zfill(3) + extension
+                fullName = os.path.join(self.statsPath, file + suffix)
+                if not os.path.exists(fullName):
+                    return fullName
+                counter += 1
+        if self.statsPath and self.scaleable:
+            self.stop = True
+            for thread in self.statsThreads:
+                thread.join()
+            fileName = getFileName()
+            with open(fileName, 'w') as f:
+                json.dump(self.stats, f)
+
+    def startStats(self, preemptable):
+        thread = ExceptionalThread(target=self._gatherStats, args=[preemptable])
+        thread.start()
+        self.statsThreads.append(thread)
+
+    def checkStats(self):
+        for thread in self.statsThreads:
+            # propagate any errors raised in the threads execution
+            thread.join(timeout=0)
+
+    def _gatherStats(self, preemptable):
+        def toDict(nodeInfo):
+            # convert NodeInfo object to dict to improve JSON output
+            return dict(memory=nodeInfo.memoryUsed,
+                        cores=nodeInfo.coresUsed,
+                        memoryTotal=nodeInfo.memoryTotal,
+                        coresTotal=nodeInfo.coresTotal,
+                        requestedCores=nodeInfo.requestedCores,
+                        requestedMemory=nodeInfo.requestedMemory,
+                        workers=nodeInfo.workers,
+                        time=time.time()  # add time stamp
+                        )
+        if self.scaleable:
+            stats = {}
+            try:
+                while not self.stop:
+                    nodeInfo = self.batchSystem.getNodes(preemptable)
+                    for nodeIP in nodeInfo.keys():
+                        nodeStats = nodeInfo[nodeIP]
+                        if nodeStats is not None:
+                            nodeStats = toDict(nodeStats)
+                            try:
+                                # if the node is already registered update the dictionary with
+                                # the newly reported stats
+                                stats[nodeIP].append(nodeStats)
+                            except KeyError:
+                                # create a new entry for the node
+                                stats[nodeIP] = [nodeStats]
+                    time.sleep(60)
+            finally:
+                threadName = 'Preemptable' if preemptable else 'Non-preemptable'
+                logger.debug('%s provisioner stats thread shut down successfully.', threadName)
+                self.stats[threadName] = stats
+        else:
+            pass
